pyproject.toml
[project]
name = "kittsay"
version = "0.1"
description = "A python API / Command-line tool for communicating with KITT from the famous, original and life changing series Knight Rider."
readme = "README.md"
authors = [{name = "Patrick Pfeiler"}]
license = "GPL-3.0-only"
requires-python = ">=3.10"
classifiers = []
dependencies = [
    "pycountry>=24.6.1",
    "pydantic-ai-slim[openai]>=1.0.1",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/ppfeiler/kittsay"
Changelog = "https://github.com/ppfeiler/kittsay/releases"
Issues = "https://github.com/ppfeiler/kittsay/issues"
CI = "https://github.com/ppfeiler/kittsay/actions"

[project.scripts]
kittsay = "kittsay.__main__:cli"

[tool.uv]
dev-dependencies = [
    "pre-commit",
    "pytest",
]

[tool.ty.rules]
unresolved-import = "ignore"
# TODO: Unable to resolve submodules: https://github.com/astral-sh/ty/issues/445
unresolved-attribute = "ignore"
# TODO: Unable to work with **kwargs: https://github.com/astral-sh/ty/issues/247
missing-argument = "ignore"
